feat: Add managedplugin  (#16)
This adds managedplugin to plugin-pb-go.

In the past we had managedsourceplugin and manageddestinationplugin in the CLI. Now as we move to SDK V3 where we have one protocol for sources and destinations we also merge the `managedplugin` and we move it to plugin-pb-go so SDK and plugin can use this capability as well (for example to spawn plugins to use them as state backends.)

This is also adds a new discovery service that uses `int` instead of strings which is quite unnecessary. 

Also, this removes `specreader` from `specs` as before we had the CLI spec coupled with over-the-wire configuration sent also as spec. Now `specs` only contains over-the-wire structs while the `cli` contains it's own `specs` and `specreader`.