feat: Implement API to get latest plugin version. (#415)
Prerequisite to fully implement: https://github.com/cloudquery/cloudquery/issues/19286

In order to let CLI users when the plugins they are using are not up to date, we need to now what the latest version is.

This PR only adds this functionality within `plugin-pb-go` within the `managedplugin` module, but it isn't yet used, nor compared to the user's plugin version.

Sample check on CLI to see if it works:

```
$ go run . test-connection /Users/mariano.gappa/Code/test-aws-sync/all.yml

Loading spec(s) from /Users/mariano.gappa/Code/test-aws-sync/all.yml
You chose version v27.23.0. Latest version is v27.23.1.
```

Also note that this implementation errors in a lot of cases, but it is not the intention to actually error to the client; since the purpose is to show a warning, we'll probably elide any warnings (it's debatable) if there's no sufficient information to show a warning.